+++ NEWS.md
@@ -4,6 +4,7 @@
 - Added caching functions to support downloading and caching the full 1991-2020 dataset
   - `cache_dir()`, `cache_remove()`
 - No longer include stations data frame in weathercan, prompt user to download always (ensures it's up to date)
+- Add progress bars for downloads (#58)
 
 
 # weathercan 0.7.8